src/service/collector/requirements.txt
kserve
python-dotenv
kfp
psycopg2-binary